requirements.txt
streamlit==1.40.2
langgraph==0.2.42
langchain==0.3.7
langchain-google-genai==2.0.5
tavily-python==0.4.13
python-dotenv==1.0.1
pydantic==2.9.2
aiohttp==3.10.11
requests==2.32.3
